Implement and integrate basic automation triggers
The update introduces basic automation triggers for two lending protocols: Aave's auto-buy and auto-sell. This consists of changes on type definitions and state management for handling these new triggers. The modification allows the use and management of these triggers strategically, creating a more streamlined process for managing lending operations. Besides, the update includes enhancements in error handling and increases code's efficiency.
