+++ helpers/triggers/setup-triggers/index.ts
@@ -1,2 +1,3 @@
 export * from './setup-aave-auto-buy'
 export * from './setup-triggers-types'
+export * from './setup-aave-auto-sell'
